Optimize permission admin pages
Also removes the service sector admin pages
