+++ permissions/admin/__init__.py
@@ -1,0 +1,9 @@
+from .general_role import GeneralRoleAdmin, GeneralRoleChoiceAdmin
+from .unit_role import UnitRoleAdmin, UnitRoleChoiceAdmin
+
+__all__ = [
+    "GeneralRoleAdmin",
+    "GeneralRoleChoiceAdmin",
+    "UnitRoleAdmin",
+    "UnitRoleChoiceAdmin",
+]
